requirements.txt
requests
python-dotenv
flask
flask_cors
flask_socketio
eventlet
